Seperated out class_singleWorker (POW thread)
